hestia/requirements.txt
asgiref==3.2.7
cffi==1.14.0
cryptography==2.8
dj-database-url==0.5.0
Django==3.0.4
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.4
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1
